Switched to Cog.commands for modularity and code organisation
