+++ .env.example
@@ -1,7 +1,7 @@
 TOKEN='YES'
-CHANNEL_ID=1
+BOT_CHANNEL_ID=1
 APPROVAL_CHANNEL_ID=1
-SERVER_ID=1
+GUILD_ID=1
 LOG_CHANNEL_ID=1
 APPROVER_ROLE_ID=1
 ONBOARDING_EXCLUSION_ROLE_ID=1
