+++ Dockerfile
@@ -1,5 +1,7 @@
 FROM python:3.12-slim
-WORKDIR /app
+
+WORKDIR /traxus
 COPY . .
 RUN pip install --no-cache-dir -r requirements.txt
-CMD ["python", "main.py"]
+
+CMD ["python", "-u", "cogs"]
