+++ docker-compose.yml
@@ -1,0 +1,5 @@
+services:
+  traxus:
+    container_name: traxus
+    build: ./
+    env_file: .env
